Solve the set of valid parentheses for n sets of parens
Created a solution for the parentheses problem, leetcode 22. Generate
Parentheses.  Took the approach of a binary analog to parens.  Since
the solution requires testing lots of different combinations, it makes
sense to increment a number and use the bits because that will cover
every possible combination of bits/parens.

Included some optimization.  You're never going to start a set of
valid parens with ")", so don't even iterate on numbers that start
with a 0 in the most significant bit.  More optimization is possible
for low cost, but this is already a decent start.  I'll improve in
future commits.

In short, establish a start and end number range based on the number
of bits needed to represent the requested number of parens.  Test each
candidate value for validity and convert the valid ones to parens.

Parentheses is neither easy to say or type.

Solves this problem:
https://leetcode.com/problems/generate-parentheses/